GH-10 Webpack config builder
