+++ package.json
@@ -92,6 +92,7 @@
     "lint-staged": "^13.2.3",
     "postcss-styled-syntax": "^0.4.0",
     "prettier": "^2.8.8",
+    "ramda": "^0.29.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-refresh": "^0.14.0",
