+++ packages/core/webpack.config.js
@@ -1,95 +1,25 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const path = require('path');
-const webpack = require('webpack');
-const TerserPlugin = require('terser-webpack-plugin');
+const createConfig = require('../../webpack.packages.base');
 
-const isProduction = process.env.NODE_ENV == 'production';
-
-const OUTPUT_PATH = path.resolve(__dirname, 'dist');
-
-const config = {
-  entry: ['./src/index.ts'],
-  output: {
-    path: OUTPUT_PATH,
-    filename: 'index.js',
-    // https://webpack.js.org/guides/output-management/#cleaning-up-the-dist-folder
-    clean: true,
-    library: {
-      name: 'SSACore',
-      type: 'umd',
-    },
-  },
-  externals: {
-    react: 'react',
-    'react-dom': 'react-dom',
-    '@emotion/core': '@emotion/core',
-    '@emotion/css': '@emotion/css',
-    '@emotion/react': '@emotion/react',
-    '@emotion/styled': '@emotion/styled',
-  },
-  // experiments: { outputModule: true },
-  devServer: {
-    open: true,
-    host: 'localhost',
-    hot: true,
-    client: {
-      logging: 'verbose',
-    },
-  },
-  devtool: isProduction ? 'source-map' : 'eval',
-  optimization: {
-    // Remove License.txt files from the ./dist/
-    // https://github.com/webpack-contrib/terser-webpack-plugin#remove-comments
-    minimize: true,
-    minimizer: [
-      new TerserPlugin({
-        terserOptions: {
-          format: {
-            comments: false,
-          },
-        },
-        extractComments: false,
-      }),
-    ],
-  },
-  plugins: [
-    new webpack.DefinePlugin({
-      'process.env': JSON.stringify(process.env),
-    }),
-  ],
-  module: {
-    rules: [
-      {
-        test: /\.(ts|tsx)$/i,
-        loader: 'babel-loader',
-        exclude: ['/node_modules/'],
-        options: {
-          configFile: '../../.babelrc.js',
-        },
-      },
-      {
-        test: /\.(eot|svg|ttf|woff|woff2|png|jpg|gif)$/i,
-        type: 'asset',
-      },
-    ],
-  },
-  resolve: {
-    extensions: ['.tsx', '.ts', '.jsx', '.js', '...'],
+module.exports = () => {
+  const currentConfig = createConfig({
+    libraryName: 'SSACore',
     alias: {
       '@components': path.resolve(__dirname, './src/components'),
       '@themes': path.resolve(__dirname, './src/themes'),
       '@styles': path.resolve(__dirname, './src/styles'),
       '@hooks': path.resolve(__dirname, './src/hooks'),
     },
-  },
-};
+    externals: {
+      react: 'react',
+      'react-dom': 'react-dom',
+      '@emotion/core': '@emotion/core',
+      '@emotion/css': '@emotion/css',
+      '@emotion/react': '@emotion/react',
+      '@emotion/styled': '@emotion/styled',
+    },
+  });
 
-module.exports = () => {
-  if (isProduction) {
-    config.mode = 'production';
-  } else {
-    config.mode = 'development';
-  }
-
-  return config;
+  return currentConfig;
 };
